Use usize instead of u64 for linux kernel FFI

+++ rust/src/constants.rs
@@ -15,30 +15,30 @@ pub enum fb_dev_op_t {
 pub use fb_dev_op_t::*;
 
 // Hardware configuration
-pub const NUM_BUSES: u32 = 1;
+pub const NUM_BUSES: usize = 1;
 
-pub const NUM_CHIPS_PER_BUS: u32 = 2;
-pub const NUM_BLOCKS_PER_CHIP: u32 = 171;
+pub const NUM_CHIPS_PER_BUS: usize = 2;
+pub const NUM_BLOCKS_PER_CHIP: usize = 171;
 /// v-layer * h-layer * multi-level degree
-pub const NUM_PAGES_PER_BLOCK: u32 = 16 * 4 * 3;
+pub const NUM_PAGES_PER_BLOCK: usize = 16 * 4 * 3;
 
 //
 // option.h
 //
 
 // Hardware configuration
-pub const CFACTOR_PERCENT: u32 = 90;
-pub const NR_LP_IN_PP: u32 = 4;
+pub const CFACTOR_PERCENT: usize = 90;
+pub const NR_LP_IN_PP: usize = 4;
 pub const LP_PAGE_MASK: u32 = 3;
 pub const LP_PAGE_SHIFT: u32 = 2;
 
-pub const NUM_CHIPS: u32 = NUM_BUSES * NUM_CHIPS_PER_BUS;
-pub const NUM_BLOCKS: u32 = NUM_CHIPS * NUM_BLOCKS_PER_CHIP;
-pub const NUM_PAGES: u32 = NUM_BLOCKS * NUM_PAGES_PER_BLOCK;
-pub const NUM_LOG_PAGES: u32 = NUM_PAGES * NR_LP_IN_PP * CFACTOR_PERCENT / 100;
+pub const NUM_CHIPS: usize = NUM_BUSES * NUM_CHIPS_PER_BUS;
+pub const NUM_BLOCKS: usize = NUM_CHIPS * NUM_BLOCKS_PER_CHIP;
+pub const NUM_PAGES: usize = NUM_BLOCKS * NUM_PAGES_PER_BLOCK;
+pub const NUM_LOG_PAGES: usize = NUM_PAGES * NR_LP_IN_PP * CFACTOR_PERCENT / 100;
 
-pub const LOGICAL_PAGE_SIZE: u32 = 4096;
-pub const PHYSICAL_PAGE_SIZE: u32 = LOGICAL_PAGE_SIZE * NR_LP_IN_PP;
+pub const LOGICAL_PAGE_SIZE: usize = 4096;
+pub const PHYSICAL_PAGE_SIZE: usize = LOGICAL_PAGE_SIZE * NR_LP_IN_PP;
 
 pub const NUM_MAX_ENTRIES_OPR_QUEUE: u32 = 4;
